update MVC and migration

+++ database/migrations/2018_01_15_025909_create_blok_table.php
@@ -16,8 +16,8 @@ class CreateBlokTable extends Migration
         Schema::create('blok', function (Blueprint $table) {
             $table->increments('id');
             $table->string('nama');
-            $table->integer('jum_lot_total');
-            $table->float('anggaran_kos');
+            $table->float('jum_lot_total',8,2);
+            $table->float('anggaran_kos',10,2);
             $table->integer('status_batal');
             $table->integer('fasa_pengambilan');
             $table->string('rujukan_jkptg');
